Move app code to package, update imports
