+++ library_registry/adobe_xml_templates.py
