+++ library_registry/emailer.py
@@ -4,7 +4,7 @@ from email.mime.text import MIMEText
 from email import charset
 import smtplib
 
-from config import CannotLoadConfiguration, CannotSendEmail
+from library_registry.config import CannotLoadConfiguration, CannotSendEmail
 
 # Set up an encoding/decoding between UTF-8 and quoted-printable.
 # Otherwise, the bodies of email messages will be encoded with base64
@@ -178,7 +178,7 @@ class Emailer:
     @classmethod
     def _sitewide_integration(cls, _db):
         """Find the ExternalIntegration for the emailer."""
-        from model import ExternalIntegration
+        from library_registry.model import ExternalIntegration
         qu = _db.query(ExternalIntegration).filter(
             ExternalIntegration.goal == cls.GOAL
         )
