+++ library_registry/geometry_loader.py
@@ -3,12 +3,12 @@ from geoalchemy2 import Geometry
 from sqlalchemy import func
 from sqlalchemy.sql.expression import cast
 
-from model import (
+from library_registry.model import (
     get_one_or_create,
     Place,
     PlaceAlias,
 )
-from util import GeometryUtility
+from library_registry.util import GeometryUtility
 
 class GeometryLoader(object):
     """Load Place objects from a NDJSON document like that generated by
