+++ library_registry/templates.py
