+++ library_registry/util/http.py
@@ -2,7 +2,7 @@ import logging
 import requests
 import urllib.parse
 from flask_babel import lazy_gettext as _
-from .problem_detail import (
+from library_registry.util.problem_detail import (
     ProblemDetail as pd,
     JSON_MEDIA_TYPE as PROBLEM_DETAIL_JSON_MEDIA_TYPE,
 )
