+++ tests/test_adobe_vendor_id.py
@@ -1,7 +1,7 @@
 import json
 
-import adobe_xml_templates as t
-from adobe_vendor_id import (
+import library_registry.adobe_xml_templates as t
+from library_registry.adobe_vendor_id import (
     AdobeSignInRequestParser,
     AdobeAccountInfoRequestParser,
     AdobeVendorIDClient,
@@ -10,14 +10,14 @@ from adobe_vendor_id import (
     VendorIDAuthenticationError,
     VendorIDServerException,
 )
-from config import Configuration
-from model import (
+from library_registry.config import Configuration
+from library_registry.model import (
     DelegatedPatronIdentifier,
     ExternalIntegration,
     create,
 )
-from util.short_client_token import ShortClientTokenEncoder
-from util.string_helpers import base64
+from library_registry.util.short_client_token import ShortClientTokenEncoder
+from library_registry.util.string_helpers import base64
 
 from . import DatabaseTest
