+++ tests/test_app.py
@@ -2,12 +2,12 @@ from io import BytesIO
 import contextlib
 import flask
 import gzip
-from app_helpers import (
+from library_registry.app_helpers import (
     compressible,
     has_library_factory,
     uses_location_factory,
 )
-from problem_details import (
+from library_registry.problem_details import (
     LIBRARY_NOT_FOUND
 )
 from .test_controller import ControllerTest
